@@ -1,5 +1,5 @@
 {
-  "name": "udagram-frontend",
+  "name": "udacity-c2-frontend",
   "version": "0.0.1",
   "author": "Ionic Framework",
   "homepage": "https://ionicframework.com/",
@@ -13,50 +13,47 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/forms": "^8.2.14",
-    "@angular/http": "^7.2.2",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
-    "@ionic-native/core": "^5.0.0",
-    "@ionic-native/splash-screen": "^5.0.0",
-    "@ionic-native/status-bar": "^5.0.0",
-    "@ionic/angular": "^4.1.0",
-    "@ionic/cli": "^6.16.3",
-    "core-js": "^2.5.4",
-    "fsevents": "^2.3.2",
-    "rxjs": "~6.5.4",
-    "zone.js": "~0.9.1"
+    "@angular/common": "^11.0.5",
+    "@angular/core": "^11.0.5",
+    "@angular/forms": "^11.0.5",
+    "@angular/platform-browser": "^11.0.5",
+    "@angular/platform-browser-dynamic": "^11.0.5",
+    "@angular/router": "^11.0.5",
+    "@ionic-native/core": "^5.9.0",
+    "@ionic-native/splash-screen": "^5.9.0",
+    "@ionic-native/status-bar": "^5.9.0",
+    "@ionic/angular": "^5.5.2",
+    "core-js": "^3.0.0",
+    "rxjs": "^6.6.3",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.12.3",
-    "@angular-devkit/build-angular": "^0.803.24",
-    "@angular-devkit/core": "~7.2.3",
-    "@angular-devkit/schematics": "~7.2.3",
-    "@angular/cli": "~8.3.25",
-    "@angular/compiler": "~8.2.14",
-    "@angular/compiler-cli": "~8.2.14",
-    "@angular/language-service": "~8.2.14",
-    "@ionic/angular-toolkit": "~1.4.0",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~10.12.0",
-    "@typescript-eslint/eslint-plugin": "^2.20.0",
-    "@typescript-eslint/parser": "^2.20.0",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.4",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
-    "ts-node": "~8.0.0",
-    "tslint": "~5.12.0",
-    "typescript": "^3.5.3"
+    "@angular-devkit/architect": "^0.1100.5",
+    "@angular-devkit/build-angular": "~0.1100.5",
+    "@angular-devkit/core": "^11.0.5",
+    "@angular-devkit/schematics": "^11.0.5",
+    "@angular/cli": "^11.0.5",
+    "@angular/compiler": "^11.0.5",
+    "@angular/compiler-cli": "^11.0.5",
+    "@angular/language-service": "^11.0.5",
+    "@ionic/angular-toolkit": "^3.0.0",
+    "@types/jasmine": "~3.6.0",
+    "@types/jasminewd2": "^2.0.3",
+    "@types/node": "^12.11.1",
+    "ajv": "^6.12.6",
+    "codelyzer": "^6.0.1",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.1.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
+    "ts-node": "^8.0.0",
+    "tslint": "~6.1.0",
+    "typescript": "~4.0.5"
   },
   "description": "An Ionic project"
 }
