@@ -1,7 +1,9 @@
 import { Injectable } from '@angular/core';
-import { HttpClient, HttpHeaders, HttpRequest, HttpEvent } from '@angular/common/http';
+import { HttpClient, HttpHeaders,  HttpErrorResponse, HttpRequest, HttpEvent, HttpEventType } from '@angular/common/http';
 import { environment } from '../../environments/environment';
-import { map } from 'rxjs/operators';
+import { Observable } from 'rxjs';
+import { FeedItem } from '../feed/models/feed-item.model';
+import { catchError, tap, map } from 'rxjs/operators';
 
 const API_HOST = environment.apiHost;
 
@@ -18,28 +20,23 @@ export class ApiService {
   constructor(private http: HttpClient) {
   }
 
-  static handleError(error: Error) {
+  handleError(error: Error) {
     alert(error.message);
   }
 
-  static extractData(res: HttpEvent<any>) {
-    const body = res;
-    return body || { };
-  }
-
   setAuthToken(token) {
-    this.httpOptions.headers = this.httpOptions.headers.append('Authorization', `jwt ${token}`);
+    this.httpOptions.headers = this.httpOptions.headers.set('Authorization', `bearer ${token}`);
     this.token = token;
   }
 
   get(endpoint): Promise<any> {
     const url = `${API_HOST}${endpoint}`;
-    const req = this.http.get(url, this.httpOptions).pipe(map(ApiService.extractData));
+    const req = this.http.get(url, this.httpOptions).pipe(map(this.extractData));
 
     return req
             .toPromise()
             .catch((e) => {
-              ApiService.handleError(e);
+              this.handleError(e);
               throw e;
             });
   }
@@ -49,7 +46,7 @@ export class ApiService {
     return this.http.post<HttpEvent<any>>(url, data, this.httpOptions)
             .toPromise()
             .catch((e) => {
-              ApiService.handleError(e);
+              this.handleError(e);
               throw e;
             });
   }
@@ -66,10 +63,16 @@ export class ApiService {
 
     return new Promise ( resolve => {
         this.http.request(req).subscribe((resp) => {
-        if (resp && (<any> resp).status && (<any> resp).status === 200) {
+        if (resp && resp.type === HttpEventType.Response && (<any> resp).status && (<any> resp).status === 200) {
           resolve(this.post(endpoint, payload));
         }
       });
     });
+  }
+
+  /// Utilities
+  private extractData(res: HttpEvent<any>) {
+    const body = res;
+    return body || { };
   }
 }
