@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { AuthMenuUserComponent } from './auth-menu-user.component';
 import { ModalController } from '@ionic/angular';
@@ -10,12 +10,13 @@ describe('AuthMenuUserPage', () => {
   let modalSpy;
   let modalCtrlSpy;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     modalSpy = jasmine.createSpyObj('Modal', ['dismiss']);
     modalCtrlSpy = jasmine.createSpyObj('ModalController', ['create']);
     modalCtrlSpy.create.and.callFake(function () {
         return modalSpy;
     });
+
 
     TestBed.configureTestingModule({
       providers: [
@@ -39,4 +40,10 @@ describe('AuthMenuUserPage', () => {
   it('should create', () => {
     expect(component).toBeTruthy();
   });
+
+  // it('#dismiss() should hide the modal', () => {
+  //     .........
+  //     expect(modalSpy.dismiss).toHaveBeenCalled(); 
+  //     .........
+  // });
 });
