@@ -1,4 +1,5 @@
-import { Component, OnInit, Input, OnDestroy } from '@angular/core';
+import { Component, OnInit, ChangeDetectionStrategy, Input, OnDestroy } from '@angular/core';
+import { environment } from '../../../environments/environment';
 import { FeedItem } from '../models/feed-item.model';
 import { FeedProviderService } from '../services/feed.provider.service';
 import { Subscription } from 'rxjs';
@@ -7,6 +8,7 @@ import { Subscription } from 'rxjs';
   selector: 'app-feed-list',
   templateUrl: './feed-list.component.html',
   styleUrls: ['./feed-list.component.scss'],
+  // changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class FeedListComponent implements OnInit, OnDestroy {
   @Input() feedItems: FeedItem[];
@@ -27,4 +29,6 @@ export class FeedListComponent implements OnInit, OnDestroy {
       subscription.unsubscribe();
     }
   }
+
+
 }
