@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { FeedUploadButtonComponent } from './feed-upload-button.component';
 
@@ -7,7 +7,7 @@ describe('FeedUploadButtonPage', () => {
   let component: FeedUploadButtonComponent;
   let fixture: ComponentFixture<FeedUploadButtonComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ FeedUploadButtonComponent ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
