@@ -1,5 +1,5 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { FeedUploadComponent } from './feed-upload.component';
 
@@ -7,7 +7,7 @@ describe('FeedUploadPage', () => {
   let component: FeedUploadComponent;
   let fixture: ComponentFixture<FeedUploadComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ FeedUploadComponent ],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
